requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
sqlalchemy==2.0.23
feedparser==6.0.10
python-dotenv==1.0.0
flask==3.0.0
schedule==1.2.0
python-dateutil==2.8.2

